*Use correct format specifiers when loading in ObjectStorage.cpp *Reorganised into alphabetical order *Return an Int32 from GetInt32 *AIAgents, UInt8 shall be read as UInt8
git-svn-id: svn://svn.aspiredev.org/AspireCore/branches/Hearthstone@1089 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/hearthstone-world/Quest.h
@@ -185,7 +185,7 @@ struct Quest
 	uint32 required_quest_and_or;
 	uint32 receive_items[4];
 	uint32 receive_itemcount[4];
-	int is_repeatable;
+	uint8 is_repeatable;
 
 	//Calculated variables (not in db)
 	uint32 count_required_mob;
